Add relation to sphinx
From: https://github.com/orgs/sphinx-doc/discussions/13331#discussioncomment-12310597

While the groups other than users are comparably small, I anticipate that they will take part overproportionally in the survey because the are more involved with the project, and they may have significantly different perspectives and answers compared to users. It therefore may be helpful to be able to filter on this when evaluating the survey.